my_ai_web_app/requirements.txt
Flask
ultralytics
Pillow
opencv-python
Werkzeug
kagglehub
